odoo-doodba-dev/skills/reasoningbank/pyproject.toml
[project]
name = "reasoningbank-skill"
version = "2.0.0"
description = "ReasoningBank - Learning memory system for Claude Code"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Letzdoo (Jerome Sonnet)"}
]

dependencies = [
    "aiosqlite>=0.19.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "ulid-py>=1.1.0",
    "anthropic>=0.39.0",
    "pyyaml>=6.0",
    "voyageai>=0.2.0",
]

[project.scripts]
rb-retrieve = "scripts.retrieve:main"
rb-judge = "scripts.judge:main"
rb-distill = "scripts.distill:main"
rb-consolidate = "scripts.consolidate:main"
rb-matts = "scripts.matts:main"
rb-pre-hook = "scripts.pre_hook:main"
rb-post-hook = "scripts.post_hook:main"
rb-init-db = "scripts.init_db:main"
rb-status = "scripts.status:main"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
